move BuildState and JobDeps up modules
